refactored where fahrt.views live. The monoliythic-giagantic-file approach was not working out
